% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cess.R
\name{cess}
\alias{cess}
\title{Calculate estimates of costs and economies of scale and scope from 25 to 200 \% levels at means}
\usage{
cess(data, outputName, priceName, controlName, model,
  vcovCL = clusterEst(model = model, cluster = unidat$unicode)$vcovCL)
}
\arguments{
\item{data}{The data used for calculating the estimates.}

\item{outputName}{A vector of strings containing the names of the independent (output) variables.}

\item{priceName}{A vector of strings containing the names of the independent (price) variables.}

\item{controlName}{A vector of strings containing the names of the control variables.}

\item{model}{The estimated model(nls class object).}

\item{vcovCL}{A variance matrix provided by clusterEst function}

\item{form}{A cost function character.}
}
\value{
Estimates of scale and scope economies including their standard errors (SE), lower
        interval(Lo) and upper interval(Hi)
}
\description{
Calculate estimates of costs and economies of scale and scope from 25 to 200 \% levels at means
}
\examples{
##Reproduce results of Zhang et al. (in press)
data(unidat)
data = unidat
library(minpack.lm)
model <- nlsLM(costFunction(costName = colnames(unidat)[3], outputName = colnames(unidat)[7:11],
priceName = colnames(unidat)[4:6], controlName = colnames(unidat)[12:24],
form = "FFCQ-M"), start = list(b0 = 600, b1 = 0, b2 = 0,
                               b3 = 0, b4 = 0, b5 = 0, b11 = 0, b22 = 0, b33 = 0, b44 = 0,
                               b55 = 0, b12 = 0, b13 = 0, b14 = 0, b15 = 0, b23 = 0, b24 = 0,
                               b25 = 0, b34 = 0, b35 = 0, b45 = 0, bp2 = 0, bp3 = 0, bz1 = 0,
                               bz2 = 0, bz3 = 0, bz4 = 0, bz5 = 0, bz6 = 0, bz7 = 0, bz8 = 0,
                               bz9 = 0, bz10 = 0, bz11 = 0, bz12 = 0, bz13 = 0), data = unidat,
                               trace = F)
vcovCL <- clusterEst(model = model , cluster = unidat$unicode)$vcovCL
cess(data=data, outputName = colnames(unidat)[7:11],priceName = colnames(unidat)[4:6],
controlName = colnames(unidat)[12:24], model=model, vcovCL=vcovCL)

}
\author{
Liang-Cheng Zhang
}
\references{
Zhang, L.-C., Worthington, A. C., & Hu, M. (in press). Cost economies in the provision of
higher education for international students: Australian evidence. Higher Education. doi: \href{http://dx.doi.org/10.1007/s10734-016-0078-9}{10.1007/s10734-016-0078-9}

Zhang, L.-C., & Worthington, A. C. (2015). Evaluating the accuracy of scope economies: comparisons among
delta method, bootstrap, and Bayesian approach. Paper presented at Australian Conference of Economists
PhD Colloquium. Retrieved from \href{http://www98.griffith.edu.au/dspace/handle/10072/69409}{http://www98.griffith.edu.au/dspace/handle/10072/69409}
}

