% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunction.R
\name{costFunction}
\alias{costFunction}
\title{Generate the right format of cost function for calculating economies of scale and scope.}
\usage{
costFunction(costName, outputName, priceName, controlName, form = c("FFCQ-M"))
}
\arguments{
\item{costName}{A character string containing the name of the dependent (cost) variable.}

\item{outputName}{A vector of strings containing the names of the independent (output) variables.}

\item{priceName}{A vector of strings containing the names of the independent (price) variables.}

\item{controlName}{A vector of strings containing the names of the control variables.}

\item{form}{A cost function character.}
}
\value{
An object of class "formula" consisting of \code{costName}, \code{outputName} and \code{priceName} based on varied functional forms.
}
\description{
Generate the right format of cost function for calculating economies of scale and scope.
}
\details{
This function returns flexible fixed cost quadratic (FFCQ) function formula based on the classification of Mayo (1984). You can find the applications in Zhang et al. (in press).
}
\examples{
##Specifiy arguments with user-identified names
costFunction(costName="c",outputName = c("y1","y2"))
costFunction(costName="c",outputName = c("y1","y2","y3"),
priceName = c("w1","w2","w3"),controlName = c("z1","z2","z3"),
form="FFCQ-M")

##Specifiy arguments with data' column names
costFunction(costName=colnames(unidat)[3],outputName =
colnames(unidat)[7:11],priceName = colnames(unidat)[4:6],
controlName = colnames(unidat)[12:24], form="FFCQ-M")
}
\author{
Liang-Cheng Zhang
}
\references{
Mayo, J. W. (1984). Multiproduct monopoly, regulation, and firm costs. Southern Economic Journal, 51(1), 208-218. doi:10.2307/1058333

Zhang, L.-C., Worthington, A. C., & Hu, M. (in press). Cost economies in the provision of higher education for international students: Australian evidence. Higher Education. doi: \href{http://dx.doi.org/10.1007/s10734-016-0078-9}{10.1007/s10734-016-0078-9}
}

